+++ src/cqportal/shieldwall/__init__.py
@@ -1,2 +1,4 @@
 from .ShieldShape import ShieldShape, shield_shape, shield_shape_mirror
 from .StraightBasic import StraightBasic
+from .Mesh import Mesh
+from .Straight import Straight